// frontend/oneCheck/src/model/transitCheckSheetDto.ts
/**
 * OneCheck.Api
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { DateOnly } from './dateOnly';
import { BlueLightEnumDto } from './blueLightEnumDto';
import { TimeOnly } from './timeOnly';


export interface TransitCheckSheetDto { 
    checkSheetId?: string;
    date?: DateOnly;
    time?: TimeOnly;
    sealNumber?: string | null;
    signature?: string | null;
    name?: string | null;
    batteryStatusInPercent?: number;
    inboundCheckAirport?: string | null;
    isContainerOperating?: boolean;
    isContainerDamaged?: boolean;
    displayedTemp?: number;
    blueLight?: BlueLightEnumDto;
    isAlarmDisplayed?: boolean;
}
export namespace TransitCheckSheetDto {
}